+++ ui/apps/platform/src/Containers/Policies/Wizard/Step5/PreviewViolations.tsx
@@ -14,6 +14,9 @@ function PreviewViolations({ alertsFromDryRun }: PreviewViolationsProps): ReactE
 
     return (
         <div>
+            <Title className="pf-u-mb-sm" headingLevel="h2">
+                Deployment results
+            </Title>
             {alertsFromDryRun.map(({ deployment, violations }, alertIndex) => {
                 /*
                  * pf-u-mb-sm separates deployment name from first list item with same spacing as subsequent list items.